@@ -8,11 +8,11 @@
     "dev:options": "npm run clean && parcel src/options/index.html -d dist/options",
     "dev:dist": "env-cmd -f .env npm run dist",
     "build:content:css": "sass src/content/sass/index.scss dist/content/styles.css",
-    "build:content:js": "parcel build src/content/index.ts -d dist/content",
+    "build:content:js": "parcel build src/content/index.ts --dist-dir dist/content --no-source-maps",
     "build:content": "npm run build:content:js && npm run build:content:css",
-    "build:options": "parcel build src/options/index.html -d dist/options --public-url .",
-    "build:bg": "parcel build src/background/index.ts -d dist/background",
-    "build:popup": "parcel build src/popup/index.html -d dist/popup",
+    "build:options": "parcel build src/options/index.html --dist-dir dist/options --public-url . --no-source-maps",
+    "build:bg": "parcel build src/background/index.ts --dist-dir dist/background --no-source-maps",
+    "build:popup": "parcel build src/popup/index.html --dist-dir dist/popup --public-url . --no-source-maps",
     "dist": "npm run clean && npm run build:content && npm run build:bg && npm run build:options && npm run build:popup && npm run cp",
     "clean": "rimraf dist && rimraf .cache",
     "cp": "copyfiles manifest.json assets/**/* dist",
@@ -21,7 +21,7 @@
   },
   "engines": {
     "node": ">=14.0.0",
-    "npm": ">=6.0.0"
+    "npm": ">=7.0.0"
   },
   "browserslist": [
     "last 10 chrome version"
@@ -31,41 +31,42 @@
     "url": "git://github.com/yurist38/linkedin-video-downloader"
   },
   "devDependencies": {
-    "@babel/core": "7.9.0",
-    "@babel/preset-env": "7.9.5",
-    "@babel/preset-typescript": "7.9.0",
-    "@types/chrome": "0.0.104",
-    "@types/jest": "25.2.1",
-    "@types/node": "13.13.2",
-    "@typescript-eslint/eslint-plugin": "2.29.0",
-    "@typescript-eslint/parser": "2.29.0",
-    "@vue/component-compiler-utils": "3.1.2",
-    "@vue/test-utils": "1.0.3",
+    "@babel/core": "7.17.5",
+    "@babel/preset-env": "7.16.11",
+    "@babel/preset-typescript": "7.16.7",
+    "@parcel/transformer-sass": "2.3.2",
+    "@parcel/transformer-vue": "2.3.2",
+    "@types/chrome": "0.0.179",
+    "@types/jest": "27.4.1",
+    "@types/node": "17.0.21",
+    "@typescript-eslint/eslint-plugin": "5.13.0",
+    "@typescript-eslint/parser": "5.13.0",
+    "@vue/component-compiler-utils": "3.3.0",
+    "@vue/test-utils": "2.0.0-rc.18",
+    "@vue/vue3-jest": "27.0.0-alpha.4",
     "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "26.0.1",
-    "copyfiles": "2.2.0",
+    "babel-jest": "27.5.1",
+    "copyfiles": "2.4.1",
+    "domexception": "4.0.0",
     "env-cmd": "10.1.0",
-    "eslint": "6.8.0",
-    "eslint-plugin-vue": "6.2.2",
+    "eslint": "8.10.0",
+    "eslint-plugin-vue": "8.5.0",
     "identity-obj-proxy": "3.0.0",
-    "jest": "26.0.1",
-    "jest-chrome": "0.7.0",
-    "parcel-bundler": "1.12.4",
+    "jest": "27.5.1",
+    "jest-chrome": "0.7.2",
+    "parcel": "2.3.2",
+    "process": "0.11.10",
     "rimraf": "3.0.2",
-    "sass": "1.26.3",
-    "ts-jest": "26.4.1",
-    "typescript": "3.8.3",
-    "vue-jest": "3.0.5",
-    "vue-template-compiler": "2.6.11"
+    "sass": "1.49.9",
+    "typescript": "4.6.2"
   },
   "dependencies": {
-    "@mdi/font": "5.9.55",
-    "@sentry/browser": "5.27.6",
-    "buefy": "0.8.20",
-    "ts-debounce": "2.0.1",
-    "vue": "2.6.11",
-    "vue-class-component": "7.2.3",
+    "@mdi/font": "6.5.95",
+    "@sentry/browser": "6.18.1",
+    "equal-vue": "0.79.3",
+    "ts-debounce": "4.0.0",
+    "vue": "3.2.31",
     "vue-hot-reload-api": "2.3.4",
-    "vue-property-decorator": "8.5.0"
+    "vue-property-decorator": "9.1.2"
   }
 }
