@@ -1,17 +1,16 @@
-module.exports = {
-  preset: 'ts-jest',
+const config = {
+  verbose: true,
   testEnvironment: 'jsdom',
   moduleFileExtensions: ['ts', 'js', 'vue'],
   rootDir: '../src',
-  testMatch: [ '**/__tests__/**/*.test.ts' ],
-  testPathIgnorePatterns: ['node_modules', '.cache', 'dist'],
-  transformIgnorePatterns: ['node_modules/'],
+  testPathIgnorePatterns: ['node_modules', '.cache', 'dist', 'assets'],
   transform: {
-    '^.+\\.ts$': 'babel-jest',
-    '^.*\\.vue$': 'vue-jest',
+    '^.+\\.ts$': ['babel-jest', { configFile: './config/babel.jest.config.js' }],
+    '^.*\\.vue$': '@vue/vue3-jest',
   },
   moduleNameMapper: {
     '^.+\\.(css|scss|sass)$': 'identity-obj-proxy',
+    '^.+\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/../config/emptyMock.js',
   },
   setupFilesAfterEnv: ['../config/jest.setup.js'],
   globals: {
@@ -32,4 +31,8 @@ module.exports = {
   },
   coverageReporters: ['text', 'html'],
   coverageDirectory: '../coverage/'
+};
+
+module.exports = async () => {
+  return config;
 };
