@@ -1,24 +1,34 @@
-import Page from '../Page';
+import Vue from 'vue';
 import Sentry from '@sentry/browser';
 
-jest.mock('../Page', () => jest.fn());
 jest.mock('@sentry/browser', () => ({
   init: jest.fn(),
   configureScope: jest.fn(jest.fn()),
 }));
+jest.mock('equal-vue', () => jest.fn());
+jest.mock('vue', () => ({
+  createApp: jest.fn(() => ({
+    use: jest.fn(() => ({
+      mount: jest.fn(),
+    }))
+  })),
+}));
+jest.mock('../popup.vue', () => 'popup');
 
-describe('content/index', () => {
+describe('popup/index', () => {
   const dsn = 'test-value';
 
-  afterEach(() => {
-    jest.resetAllMocks();
+  beforeAll(() => {
+    const appEl = document.createElement('div');
+    appEl.id = 'popup';
+    document.body.appendChild(appEl);
   });
 
-  it('Should create an instance of Page class', () => {
-    process.env.SENTRY_DSN = '';
+  it('Should create an instance of Vue', () => {
     jest.isolateModules(() => {
       require('../index');
-      expect(Page).toHaveBeenCalledTimes(1);
+      expect(Vue.createApp).toHaveBeenCalledTimes(1);
+      expect(Vue.createApp).toHaveBeenCalledWith('popup');
     });
   });
 
@@ -39,19 +49,14 @@ describe('content/index', () => {
       expect(Sentry.init).toHaveBeenCalledWith({
         dsn,
         release: `linkedin-video-downloader-ext@${process.env.npm_package_version}`,
-        ignoreErrors: [
-          'ResizeObserver loop limit exceeded',
-          't8 is not defined',
-          'Extension context invalidated'
-        ],
       });
       expect(Sentry.configureScope).toHaveBeenCalledTimes(1);
-      const cb = Sentry.configureScope.mock.calls[0][0];
+      const cb = (Sentry.configureScope as jest.Mock).mock.calls[0][0];
       const setTag = jest.fn();
       const scope = { setTag };
       cb(scope as unknown as Sentry.Scope);
       expect(setTag).toHaveBeenCalledTimes(1);
-      expect(setTag).toHaveBeenCalledWith('app', 'content');
+      expect(setTag).toHaveBeenCalledWith('app', 'popup');
     });
   });
 });
